refactor(level_editor): Move stuff to multiple modules (✨MVC✨)
